Base flipbook framework?
Here is the flipbook draft I have currently from a youtube tutorial! I think the result is a bit similar to the design draft in Figma so I hope we can work with this and that theres no major issues with it! Here is the link to the tutorial I followed also: https://youtu.be/bX7Igy27DSE?si=SRv3fRTt2N7tv7y6